+++ workspaces/lightspeed/plugins/lightspeed/CHANGELOG.md
@@ -1,5 +1,18 @@
 ## @red-hat-developer-hub/backstage-plugin-lightspeed
 
+## 2.6.6
+
+### Patch Changes
+
+- c6824df: fixed notebooks empty state backgroundColor
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-common@2.6.6
+
+## 2.6.5
+
+### Patch Changes
+
+- @red-hat-developer-hub/backstage-plugin-lightspeed-common@2.6.5
+
 ## 2.6.4
 
 ### Patch Changes