+++ workspaces/orchestrator/plugins/orchestrator-form-widgets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets
 
+## 1.10.5
+
+### Patch Changes
+
+- f47b22f: Replace values in ActiveMultiSelect when clearOnRetrigger is enabled
+- 646c581: Fix type errors in orchestrator form widgets
+
 ## 1.10.4
 
 ### Patch Changes